requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via starlette
click==8.2.1
    # via uvicorn
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fastapi==0.115.12
    # via -r requirements.in
fonttools==4.58.0
    # via matplotlib
h11==0.16.0
    # via uvicorn
idna==3.10
    # via anyio
kiwisolver==1.4.8
    # via matplotlib
matplotlib==3.10.3
    # via -r requirements.in
noise==1.2.2
    # via -r requirements.in
numpy==2.2.6
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
packaging==25.0
    # via matplotlib
pillow==11.2.1
    # via
    #   -r requirements.in
    #   matplotlib
pydantic==2.11.4
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
pygame==2.6.1
    # via -r requirements.in
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.46.2
    # via fastapi
typing-extensions==4.13.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
uvicorn==0.34.2
    # via -r requirements.in
